+++ src/services/typeAccount.js
@@ -27,3 +27,27 @@ export const newTypeAccount = async (values) => {
     })
   return response
 }
+
+export const getTypeAccount = async (query) => {
+  const storeObject = store.getState()
+
+  const headers = {
+    token: storeObject.auth.token,
+    username: storeObject.auth.username,
+  }
+
+  let response = {}
+
+  await axios.get(`${BACKEND_URL}/api/typeAccount`, { headers: headers, params: { query } }).then(
+    resp => {
+      response = resp
+    }
+  ).catch((error) => {
+    if (error.response) {
+      response = error.response
+    } else {
+      console.log('Error', error.message);
+    }
+  })
+  return response
+}
