+++ src/services/usuario.js
@@ -2,31 +2,6 @@ import axios from 'axios'
 import { BACKEND_URL } from './var'
 import { store } from '../App'
 
-
-export const getTypeAccount = async (query) => {
-  const storeObject = store.getState()
-
-  const headers = {
-    token: storeObject.auth.token,
-    username: storeObject.auth.username,
-  }
-
-  let response = {}
-
-  await axios.get(`${BACKEND_URL}/api/typeAccount`, { headers: headers, params: { query } }).then(
-    resp => {
-      response = resp
-    }
-  ).catch((error) => {
-    if (error.response) {
-      response = error.response
-    } else {
-      console.log('Error', error.message);
-    }
-  })
-  return response
-}
-
 export const getUsers = async (query) => {
   const storeObject = store.getState()
 
@@ -49,7 +24,7 @@ export const getUsers = async (query) => {
     }
   })
   return response
-} 
+}
 
 export const getResourcesByTypeAccount = async (query) => {
   const storeObject = store.getState()
@@ -73,7 +48,7 @@ export const getResourcesByTypeAccount = async (query) => {
     }
   })
   return response
-} 
+}
 
 export const NovoUsuarioService = async (values) => {
   const storeObject = store.getState()
